+++ api/helpers/songs/queue-pending.js
@@ -1,4 +1,4 @@
-/* global sails, Songs */
+/* global sails, Songs, Promise */
 
 module.exports = {
 
@@ -21,7 +21,9 @@ module.exports = {
         if (Songs.pending.length > 0)
         {
             var maps = Songs.pending.map(async (track, index) => {
-                await sails.helpers.rest.cmd('LoadTrackToTop', track, 10000, inputs.queue);
+                await sails.helpers.rest.cmd('LoadTrackToTop', track, 10000);
+                if (inputs.queue)
+                    await sails.helpers.rest.checkQueue(track);
                 //wait.for.time(1);
                 delete Songs.pending[index];
                 return true;